+++ src/main/java/com/my/blog/base/service/single/IBaseService.java
@@ -8,9 +8,9 @@ public interface IBaseService<T, E> {
 
 	long countByExample(E e);
 
-	int deleteByExample(E e);
+	long deleteByExample(E e);
 
-	int deleteByPrimaryKey(T t);
+	long deleteByPrimaryKey(T t);
 
 	long insert(T t);
